backend/tennis/requirements.copy.txt
anyascii==0.3.2
asgiref==3.8.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.7.1
Django==4.2.13
django-cors-headers==4.3.1
django-filter==24.2
django-modelcluster==6.3
django-oauth-toolkit==2.4.0
django-permissionedforms==0.1
django-taggit==5.0.1
django-treebeard==4.7.1
djangorestframework==3.15.1
draftjs-exporter==5.0.0
et-xmlfile==1.1.0
filetype==1.2.0
idna==3.7
jwcrypto==1.5.6
l18n==2021.3
laces==0.1.1
oauthlib==3.2.2
openpyxl==3.1.2
pillow==10.3.0
pillow-heif==0.16.0
psycopg==3.1.19
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests==2.32.2
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
soupsieve==2.5
sqlparse==0.5.0
telepath==0.3.1
typing-extensions==4.11.0
urllib3==2.2.1
wagtail==6.1.1
willow==1.8.0



azure-ai-ml==1.18.0
azure-common==1.1.28
azure-core==1.30.2
azure-graphrbac==0.61.1
azure-identity==1.17.1
azure-mgmt-authorization==4.0.0
azure-mgmt-containerregistry==10.3.0
azure-mgmt-core==1.4.0
azure-mgmt-keyvault==10.3.1
azure-mgmt-network==26.0.0
azure-mgmt-resource==23.1.1
azure-mgmt-storage==21.2.1
azure-storage-blob==12.21.0
azure-storage-file-datalake==12.16.0
azure-storage-file-share==12.17.0
azureml-core==1.56.0
azureml-dataprep==5.2.0
azureml-dataprep-native==41.0.0
azureml-dataprep-rslex==2.23.1
msrestazure==0.6.4
opencensus-ext-azure==1.1.13

pandas